Revised on Nov 22

+++ personal directory/hyhan/Project1/Project1/mbti.cpp
@@ -2,6 +2,6 @@
 
 #include <iostream>
 
-int Student::match_mbti() {
+double Student::match_mbti() {
 	return 0;
 }
